Pulled out everything I wanted
