+++ webAO/client/isAudio.ts
@@ -1,0 +1,6 @@
+export const isAudio = (trackname: string) => {
+    const audioEndings = [".wav", ".mp3", ".ogg", ".opus"];
+    return (
+        audioEndings.filter((ending) => trackname.endsWith(ending)).length === 1
+    );
+}
